refactor: merged RasterLegendEdit to RasterLegend to simplify
