+++ .changeset/afraid-flowers-reflect.md
@@ -19,3 +19,4 @@
 - fix: removed type from LineTypes
 - fix: fixed bug of not updating sas token for PMTiles URL in saved style.
 - refactor: merged all vector sub components into VectorLegend to simplify
+- refactor: merged RasterLegendEdit to RasterLegend to simplify
